### 2020-05-22 #2217 @whedon
[PRE REVIEW]: BluVision Macro - a software for automated powdery mildew and rust disease quantification on detached leaves.
**Submitting author:** @snowformatics (<a href="http://orcid.org/0000-0003-0536-835X">Stefanie Lueck</a>)
**Repository:** <a href="https://github.com/snowformatics/macrobot" target ="_blank">https://github.com/snowformatics/macrobot</a>
**Version:** 0.4.6
