### 2020-05-22 #2217 @snowformatics
> Failed to discover a valid open source license.

I copied the license file in the main directory, hope that's the right place.
