+++ paper/paper.bib
@@ -88,7 +88,7 @@ pages = "30 - 42",
 year = "2009",
 note = "Linking Genes to Brain Function in Health and Disease",
 issn = "0306-4522",
-doi = "https://doi.org/10.1016/j.neuroscience.2009.01.027",
+doi = "10.1016/j.neuroscience.2009.01.027",
 url = "http://www.sciencedirect.com/science/article/pii/S0306452209000487",
 author = "R.M. Bilder and F.W. Sabb and T.D. Cannon and E.D. London and J.D. Jentsch and D. Stott Parker and R.A. Poldrack and C. Evans and N.B. Freimer",
 keywords = "phenotype, genetics, genomics, informatics, cognition, psychiatry",
