### 2020-05-22 #2217 @mjsottile
:+1: @arfon Sure!  Actually, this is in my area - image processing is one of my research areas and I've worked on problems similar to the one that this submission focuses on.
