### 2020-05-26 snowformatics/macrobot#1 @AoifeHughes
JOSS Review Comments
Re: [https://github.com/openjournals/joss-reviews/issues/2259](https://github.com/openjournals/joss-reviews/issues/2259) I'll be putting review comments here.

I'll start the full review shortly, but firstly I wanted to make you aware that some images in your readme.md are broken e.g. [https://github.com/snowformatics/GSOC/blob/master/Slide1.png](https://github.com/snowformatics/GSOC/blob/master/Slide1.png). If you could check through and fix these it'll save time during the review process.

Thanks for the submission and I look forward to reviewing this.
