### 2020-05-26 snowformatics/macrobot#1 @snowformatics
Thanks, I fixed the images should be fine now...

Looking forward to your review :)
