+++ paper/paper.md
@@ -37,7 +37,7 @@ Powdery mildews and rusts are in the Top 10 of the major fungal pathogens in pla
 
 Crop protection against pathogens is mostly provided by the application of chemical agents (Popp et al. 2013), however, many of them have detrimental effects on non-target species (Stanley and Preetha, 2016).  Therefore, the trend is towards reducing the pesticide application and development of alternative and integrated protection methods (Mzoughi, 2011). The most sustainable method for crop protection appears to be the use of natural genetic resources and breeding for disease resistance (Gómez et al. 2009). Plant breeding is as old as the domestication of the first agricultural plants (>10 000 years) (Zender 2011). The “Green revolution” (Hazell, 2009) of the 1950-1960 and the more recent “omics” revolution (genomics, proteomics or metabolomics, etc.) (Yadav, 2007) introduced many new approaches and technologies but the observation methods of the breeders remained mostly unchanged. Recently, new high-throughput observation methods for Phenomics (Bilder et al., 2009) were introduced, thus providing the fundament for another breeding revolution. However, the powdery mildews and rusts are, as the majority of the plant pathogens, microscopic organisms in the initial and most critical stages of the infection process. Phenotyping of these early stages was significantly held back by the lack of technology for high-throughput phenotyping on a macroscopic and microscopic scale. 
 
-To solve this problem, we have developed the BluVision Macro framework aimed to allow strictly quantitative assessment of disease and host responses on a macroscopic level. The system consists of a hardware part – the Macrobot (Lück et al., 2020) - a multimodal imaging station and robotized sample magazine/loader, and the BluVision Macro software, described in this article. The system is designed to work with samples placed in standard containers, so-called microtiter plates (MTP), which are well-established standard in biology and medicine. The loading of the MTPs to the imaging station and the image acquisition is fully automated. The system uses a 14-bit monochrome camera (Thorlabs 8050M-GE-TE) at a resolution of 3296×2472 px. The illumination is based on narrow bandwidth isotropic LED light sources (Metaphase Exolight-ISO-14-XXX-U) with 365nm (UV), 470nm (blue), 530nm (green) and 625nm (red) peak wavelength. For each plate monochrome images in all illumination, wavelengths are acquired separately and stored in 16-bit TIFF image files. Additionally, a background illumination image is taken and used for the separation of the foreground and background \autoref{fig:example}.
+To solve this problem, we have developed the BluVision Macro framework aimed to allow strictly quantitative assessment of disease and host responses on a macroscopic level. The system consists of a hardware part – the Macrobot (Lück et al., 2020) - a multimodal imaging station and robotized sample magazine/loader, and the BluVision Macro software, described in this article. The system is designed to work with samples placed in standard containers, so-called microtiter plates (MTP), which are well-established standard in biology and medicine. The loading of the MTPs to the imaging station and the image acquisition is fully automated. The system uses a 14-bit monochrome camera (Thorlabs 8050M-GE-TE) at a resolution of 3296×2472 px. The illumination is based on narrow bandwidth isotropic LED light sources (Metaphase Exolight-ISO-14-XXX-U) with 365nm (UV), 470nm (blue), 530nm (green) and 625nm (red) peak wavelength. For each plate monochrome images in all illumination, wavelengths are acquired separately and stored in 16-bit TIFF image files. Additionally, a background illumination image is taken and used for the separation of the foreground and background {fig:example}.
 
 ![Macrobot pipline.\label{fig:example}](Macrobot_Figure1.png)
 
