+++ paper/paper.bib
@@ -29,6 +29,28 @@
  doi = {10.7717/peerj.453}
 }
 
+@article{RN14,
+   author = {Douchkov, Dimitar and Lück, Stefanie and Johrde, Annika and Nowara, Daniela and Himmelbach, Axel and Rajaraman, Jeyaraman and Stein, Nils and Sharma, Rajiv and Kilian, Benjamin and Schweizer, Patrick},
+   title = {Discovery of genes affecting resistance of barley to adapted and non-adapted powdery mildew fungi},
+   journal = {Genome Biology},
+   volume = {15},
+   number = {12},
+   pages = {518},
+   ISSN = {1474-760X},
+   DOI = {10.1186/s13059-014-0518-8},
+   url = {https://doi.org/10.1186/s13059-014-0518-8},
+   year = {2014},
+   type = {Journal Article}
+}
+
+@book{book,
+author = {Duveiller, Etienne and Singh, P.K. and Mezzalama, Monica and Singh, Ravi and Dababat, Abdelfattah},
+year = {2012},
+month = {01},
+pages = {},
+title = {Wheat Diseases and Pests: A Guide for Field Identification (2nd Edition)}
+}
+
 @article{doi:10.1109/MCSE.2011.37,
 author = {Walt,Stéfan van der  and Colbert,S. Chris  and Varoquaux,Gaël },
 title = {The NumPy Array: A Structure for Efficient Numerical Computation},
@@ -45,17 +67,15 @@ URL = {
 },
 eprint = { 
         https://aip.scitation.org/doi/pdf/10.1109/MCSE.2011.37}}
-
-
-@article{opencv_library,
-    author = {Bradski, G.},
-    citeulike-article-id = {2236121},
-    journal = {Dr. Dobb's Journal of Software Tools},
-    keywords = {bibtex-import},
-    posted-at = {2008-01-15 19:21:54},
-    priority = {4},
-    title = {{The OpenCV Library}},
-    year = {2000}
+	
+@article{article2,
+author = {J., Huerta and Singh, Ravi and Germán, Silvia and McAllum, B.D. and Park, R.F. and W.Q.Chen, and Bhardwaj, S.C. and Goyeau, and},
+year = {2011},
+month = {01},
+pages = {143-160},
+title = {Global status of wheat leaf rust caused by Puccinia triticina},
+volume = {179},
+journal = {Euphytica}
 }
 
 @article{article,
@@ -64,19 +84,6 @@ year = {2009},
 month = {01},
 pages = {},
 title = {The Asian Green Revolution}
-}
-
-@article{RN13,
-   author = {Yadav, S. P.},
-   title = {The wholeness in suffix -omics, -omes, and the word om},
-   journal = {J Biomol Tech},
-   volume = {18},
-   number = {5},
-   pages = {277},
-   ISSN = {1524-0215 (Print)
-1524-0215},
-   year = {2007},
-   type = {Journal Article}
 }
 
 @article{BILDER200930,
@@ -95,28 +102,17 @@ keywords = "phenotype, genetics, genomics, informatics, cognition, psychiatry",
[…]
