+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article {Lueck2020.03.16.993451,
+@article {2020,
 	author = {Lueck, Stefanie and Strickert, Marc and Lorbeer, Maximilian and Melchert, Friedrich and Backhaus, Andreas and Kilias, David and Seiffert, Udo and Schweizer, Patrick and Djamei, Armin and Douchkov, Dimitar},
 	title = {{\textquoteleft}Macrobot{\textquoteright}{\textendash}an automated segmentation-based system for powdery mildew disease quantification},
 	elocation-id = {2020.03.16.993451},
@@ -11,7 +11,7 @@
 	journal = {bioRxiv}
 }
 
-@article{scikit-image,
+@article{2014,
  title = {scikit-image: image processing in {P}ython},
  author = {van der Walt, {S}t\'efan and {S}ch\"onberger, {J}ohannes {L}. and
            {Nunez-Iglesias}, {J}uan and {B}oulogne, {F}ran\c{c}ois and {W}arner,
@@ -29,7 +29,7 @@
  doi = {10.7717/peerj.453}
 }
 
-@article{RN14,
+@article{2014_2,
    author = {Douchkov, Dimitar and Lück, Stefanie and Johrde, Annika and Nowara, Daniela and Himmelbach, Axel and Rajaraman, Jeyaraman and Stein, Nils and Sharma, Rajiv and Kilian, Benjamin and Schweizer, Patrick},
    title = {Discovery of genes affecting resistance of barley to adapted and non-adapted powdery mildew fungi},
    journal = {Genome Biology},
@@ -43,7 +43,7 @@
    type = {Journal Article}
 }
 
-@book{book,
+@book{2012,
 author = {Duveiller, Etienne and Singh, P.K. and Mezzalama, Monica and Singh, Ravi and Dababat, Abdelfattah},
 year = {2012},
 month = {01},
@@ -51,7 +51,7 @@ pages = {},
 title = {Wheat Diseases and Pests: A Guide for Field Identification (2nd Edition)}
 }
 
-@article{doi:10.1109/MCSE.2011.37,
+@article{2011,
 author = {Walt,Stéfan van der  and Colbert,S. Chris  and Varoquaux,Gaël },
 title = {The NumPy Array: A Structure for Efficient Numerical Computation},
 journal = {Computing in Science \& Engineering},
@@ -68,7 +68,7 @@ URL = {
 eprint = { 
         https://aip.scitation.org/doi/pdf/10.1109/MCSE.2011.37}}
 	
-@article{article2,
+@article{2011_2,
 author = {J., Huerta and Singh, Ravi and Germán, Silvia and McAllum, B.D. and Park, R.F. and W.Q.Chen, and Bhardwaj, S.C. and Goyeau, and},
 year = {2011},
 month = {01},
@@ -78,7 +78,7 @@ volume = {179},
 journal = {Euphytica}
 }
 
-@article{article,
+@article{2009,
 author = {Hazell, Peter},
 year = {2009},
 month = {01},
@@ -86,7 +86,7 @@ pages = {},
 title = {The Asian Green Revolution}
 }
 
-@article{BILDER200930,
+@article{2009_2,
 title = "Phenomics: the systematic study of phenotypes on a genome-wide scale",
 journal = "Neuroscience",
 volume = "164",
@@ -102,7 +102,7 @@ keywords = "phenotype, genetics, genomics, informatics, cognition, psychiatry",
[…]
