### 2020-06-05 3e92b4d
Update paper.md
