+++ paper/paper.bib
@@ -11,6 +11,38 @@
 	journal = {bioRxiv}
 }
 
+@article{2018,
+author = {Muhammad Azhar Nadeem and Muhammad Amjad Nawaz and Muhammad Qasim Shahid and Yıldız Doğan and Gonul Comertpay and Mehtap Yıldız and Rüştü Hatipoğlu and Fiaz Ahmad and Ahmad Alsaleh and Nitin Labhane and Hakan Özkan and Gyuhwa Chung and Faheem Shehzad Baloch},
+title = {DNA molecular markers in plant breeding: current status and recent advancements in genomic selection and genome editing},
+journal = {Biotechnology \& Biotechnological Equipment},
+volume = {32},
+number = {2},
+pages = {261-285},
+year  = {2018},
+publisher = {Taylor & Francis},
+doi = {10.1080/13102818.2017.1400401},
+
+URL = { 
+        https://doi.org/10.1080/13102818.2017.1400401
+    
+},
+eprint = { 
+        https://doi.org/10.1080/13102818.2017.1400401
+    
+}
+
+}
+
+@book{2016,
+author = {Stanley, Johnson and Gnanadhas, Preetha},
+year = {2016},
+month = {01},
+pages = {1-502},
+title = {Pesticide toxicity to non-target organisms: Exposure, toxicity and risk assessment methodologies},
+journal = {Pesticide Toxicity to Non-target Organisms: Exposure, Toxicity and Risk Assessment Methodologies},
+doi = {10.1007/978-94-017-7752-0}
+}
+
 @article{2014,
  title = {scikit-image: image processing in {P}ython},
  author = {van der Walt, {S}t\'efan and {S}ch\"onberger, {J}ohannes {L}. and
@@ -43,12 +75,40 @@
    type = {Journal Article}
 }
 
+@article{2013,
+   author = {Popp, József and Pető, Károly and Nagy, János},
+   title = {Pesticide productivity and food security. A review},
+   journal = {Agronomy for Sustainable Development},
+   volume = {33},
+   number = {1},
+   pages = {243-255},
+   ISSN = {1773-0155},
+   DOI = {10.1007/s13593-012-0105-x},
+   url = {https://doi.org/10.1007/s13593-012-0105-x},
+   year = {2013},
+   type = {Journal Article}
+}
+
 @book{2012,
 author = {Duveiller, Etienne and Singh, P.K. and Mezzalama, Monica and Singh, Ravi and Dababat, Abdelfattah},
 year = {2012},
 month = {01},
 pages = {},
 title = {Wheat Diseases and Pests: A Guide for Field Identification (2nd Edition)}
+}
+
+@article{2012_2,
+author = {DEAN, RALPH and VAN KAN, JAN A. L. and PRETORIUS, ZACHARIAS A. and HAMMOND-KOSACK, KIM E. and DI PIETRO, ANTONIO and SPANU, PIETRO D. and RUDD, JASON J. and DICKMAN, MARTY and KAHMANN, REGINE and ELLIS, JEFF and FOSTER, GARY D.},
+title = {The Top 10 fungal pathogens in molecular plant pathology},
+journal = {Molecular Plant Pathology},
+volume = {13},
+number = {4},
+pages = {414-430},
+doi = {10.1111/j.1364-3703.2011.00783.x},
+url = {https://bsppjournals.onlinelibrary.wiley.com/doi/abs/10.1111/j.1364-3703.2011.00783.x},
+eprint = {https://bsppjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1364-3703.2011.00783.x},
[…]
