+++ paper/paper.bib
@@ -153,6 +153,28 @@ keywords = "Integrated protection, Organic farming, Social and moral concerns",
 abstract = "We investigate empirically the role of moral and social concerns in farmers' decision to adopt integrated crop protection (IP) and organic farming (OF). A survey questionnaire has been sent to 1286 fruit-growers and vegetable producers located in the French areas of Alpes de Haute Provence, Hautes-Alpes and Vaucluse. Analysis of individual responses (N=243) shows that, although economic concerns play a strong role, a significant number of respondents give high importance to moral and social ones. We also examine how these considerations matter according to different crop protection strategies, that is, conventional farming, IP and OF. Using a multinomial logistic regression, we find that (1) social concerns (e.g., showing to others one's environmental commitment) drive both IP and OF adoption, (2) moral concerns (e.g., do not feel guilty about one's choices) increase the probability of organic farming adoption only, and (3) farmers who give high importance to economic concerns (e.g., cutting production costs) are less likely to adopt OF."
 }
 
+@article{2011_4,
+author = {Zeder, Melinda A.},
+title = {The Origins of Agriculture in the Near East},
+journal = {Current Anthropology},
+volume = {52},
+number = {S4},
+pages = {S221-S235},
+year = {2011},
+doi = {10.1086/659307},
+
+URL = { 
+        https://doi.org/10.1086/659307
+    
+},
+eprint = { 
+        https://doi.org/10.1086/659307
+    
+}
+,
+    abstract = { The emerging picture of plant and animal domestication and agricultural origins in the Near East is dramatically different from that drawn 16 years ago in a landmark article by Bar-Yosef and Meadow. While in 1995 there appeared to have been at least a 1,500-year gap between plant and animal domestication, it now seems that both occurred at roughly the same time, with initial management of morphologically wild future plant and animal domesticates reaching back to at least 11,500 cal BP, if not earlier. A focus on the southern Levant as the core area for crop domestication and diffusion has been replaced by a more pluralistic view that sees domestication of various crops and livestock occurring, sometimes multiple times in the same species, across the entire region. Morphological change can no longer be held to be a leading-edge indicator of domestication. Instead, it appears that a long period of increasingly intensive human management preceded the manifestation of archaeologically detectable morphological change in managed crops and livestock. Agriculture in the Near East arose in the context of broad-based systematic human efforts at modifying local environments and biotic communities to encourage plant and animal resources of economic interest. This process took place across the entire Fertile Crescent during a period of dramatic post-Pleistocene climate and environmental change with considerable regional variation in the scope and intensity of these activities as well as in the range of resources being manipulated. }
+}
+
 @article{2009,
 author = {Hazell, Peter},
 year = {2009},
