+++ paper/paper.md
@@ -33,9 +33,9 @@ bibliography: paper.bib
  
 # Summary
 
-Powdery mildews and rusts are in the Top 10 of the major fungal pathogens in plant pathology (Dean et al, 2012, Huerta-Espino et al. 2011; Wellings 2011). The effect of powdery mildews on crop yields can amount to 40% of harvested grain [@2001]. Besides being an agriculturally important pathogen, the powdery mildews of wheat and barley are important models for studying the plant-pathogen interactions [@2014_2]. Wheat leaf rust and stripe rust are two other fungal pathogens, frequently causing epidemics with up to 70% yield losses [@2005] [@2011_2] in combination with a decreased grain quality (Prescott et al. 1986). 
+Powdery mildews and rusts are in the Top 10 of the major fungal pathogens in plant pathology [@2012_2] [@2011_2]. The effect of powdery mildews on crop yields can amount to 40% of harvested grain [@2001]. Besides being an agriculturally important pathogen, the powdery mildews of wheat and barley are important models for studying the plant-pathogen interactions [@2014_2]. Wheat leaf rust and stripe rust are two other fungal pathogens, frequently causing epidemics with up to 70% yield losses [@2005] [@2011_2] in combination with a decreased grain quality [@2012]. 
 
-Crop protection against pathogens is mostly provided by the application of chemical agents (Popp et al. 2013), however, many of them have detrimental effects on non-target species (Stanley and Preetha, 2016).  Therefore, the trend is towards reducing the pesticide application and development of alternative and integrated protection methods (Mzoughi, 2011). The most sustainable method for crop protection appears to be the use of natural genetic resources and breeding for disease resistance (Gómez et al. 2009). Plant breeding is as old as the domestication of the first agricultural plants (>10 000 years) (Zender 2011). The “Green revolution” [@2009] of the 1950-1960 and the more recent “omics” revolution (genomics, proteomics or metabolomics, etc.) [@2007] introduced many new approaches and technologies but the observation methods of the breeders remained mostly unchanged. Recently, new high-throughput observation methods for Phenomics [@2009_2] were introduced, thus providing the fundament for another breeding revolution. However, the powdery mildews and rusts are, as the majority of the plant pathogens, microscopic organisms in the initial and most critical stages of the infection process. Phenotyping of these early stages was significantly held back by the lack of technology for high-throughput phenotyping on a macroscopic and microscopic scale. 
+Crop protection against pathogens is mostly provided by the application of chemical agents [@2013], however, many of them have detrimental effects on non-target species [@2016].  Therefore, the trend is towards reducing the pesticide application and development of alternative and integrated protection methods [@2011_3]. The most sustainable method for crop protection appears to be the use of natural genetic resources and breeding for disease resistance [@2009_3]. Plant breeding is as old as the domestication of the first agricultural plants (>10 000 years) [@2011_4]. The “Green revolution” [@2009] of the 1950-1960 and the more recent “omics” revolution (genomics, proteomics or metabolomics, etc.) [@2007] introduced many new approaches and technologies but the observation methods of the breeders remained mostly unchanged. Recently, new high-throughput observation methods for Phenomics [@2009_2] were introduced, thus providing the fundament for another breeding revolution. However, the powdery mildews and rusts are, as the majority of the plant pathogens, microscopic organisms in the initial and most critical stages of the infection process. Phenotyping of these early stages was significantly held back by the lack of technology for high-throughput phenotyping on a macroscopic and microscopic scale. 
 
[…]
