+++ paper/paper.bib
@@ -98,7 +98,7 @@ title = {Wheat Diseases and Pests: A Guide for Field Identification (2nd Edition
 }
 
 @article{2012_2,
-author = {DEAN, RALPH and VAN KAN, JAN A. L. and PRETORIUS, ZACHARIAS A. and HAMMOND-KOSACK, KIM E. and DI PIETRO, ANTONIO and SPANU, PIETRO D. and RUDD, JASON J. and DICKMAN, MARTY and KAHMANN, REGINE and ELLIS, JEFF and FOSTER, GARY D.},
+author = {Dean, RALPH and VAN KAN, JAN A. L. and PRETORIUS, ZACHARIAS A. and HAMMOND-KOSACK, KIM E. and DI PIETRO, ANTONIO and SPANU, PIETRO D. and RUDD, JASON J. and DICKMAN, MARTY and KAHMANN, REGINE and ELLIS, JEFF and FOSTER, GARY D.},
 title = {The Top 10 fungal pathogens in molecular plant pathology},
 journal = {Molecular Plant Pathology},
 volume = {13},
@@ -129,7 +129,7 @@ eprint = {
         https://aip.scitation.org/doi/pdf/10.1109/MCSE.2011.37}}
 	
 @article{2011_2,
-author = {J., Huerta and Singh, Ravi and Germán, Silvia and McAllum, B.D. and Park, R.F. and W.Q.Chen, and Bhardwaj, S.C. and Goyeau, and},
+author = {Huerta and Singh, Ravi and Germán, Silvia and McAllum, B.D. and Park, R.F. and W.Q.Chen, and Bhardwaj, S.C. and Goyeau, and},
 year = {2011},
 month = {01},
 pages = {143-160},
