### 2020-06-05 8f74f6b
Update paper.md
