### 2020-06-05 snowformatics/macrobot#1 @snowformatics
**Dear Nathan. Thank you very much for your critical reading and valuable comments. Your input is
highly appreciated.**

> # Paper Comments
> I'd like to prefix these comments by saying that they are general suggestions that I, as the reader, feel would help improve readability and clarity. If they seem blunt, this is not the intent :)
>
> * I think a reference to a review article for the first few sentences would be good.

**[Done]**
> * Use of the word 'probably'. Personally I feel this isn't a good word to use when making a scientific point, particularly when not backed up with references.

**We have rephrased the sentence and added references but, in our opinion, it is too early for making definitive statements, which is the most sustainable approach for crop protection. Therefore, we left some ambiguity in the sentence.**

> * "The plant breeding is as old as the domestication ..." This sentence doesn't make sense. You could remove "The" and it would read better.

**[Done]**

> * A more general comment would be to consider spacing out information in paragraphs. The first page is essentially a single paragraph. It may work better if you made a single paragraph introduction to pathogens being an issue and how plant breeding attempts to tackle problems then a second paragraph which states how your approach feeds into this.

**Good point! [Done]**

> * The phrasing of "a combined term
>   that refers to the fields of study in biology ending in -omics," is superfluous. It could simply be "(genomics, proteomics, metabolomics etc.)"

**[Done]**

> * Again, I would consider restructuring. You talk of the breeding for 10k years, jump to omics then back to breeders not using them. It may work better if mentioned how most breeding hasn't evolved since domestication and that your method helps to bring in 21st century technology.

**[Done]**

> * "Introduction ~cing~ of ~the~ so-called".

**[Done]**

> * Overuse of "so-called". Possibly not required at all.

**[Done]**

> * "... significantly withdrawn" I don't think this word means what you want to say. Maybe "arrested" or plainly state "Early phenotyping was held back by the difficulty in developing high-throughput pipelines for both macroscopic and microscopic data".

**Sorry, we probably wanted to write “withhold” or so...It is corrected.**

> * "To contribute to" -> "To solve these problems". Use of "to" twice doesn't read well.

**[Done]**

> * The introduction of your method could be a new paragraph. It would really help the reader jump straight to reading about what you have done. As currently it is hidden amongst background information.

**[Done]**

> * After you introduce your method you jump back to the effects of powdery mildew. As I mentioned previously, this could be part of the introduction and justification of this work.

**[Done]**

> * Figure 1. A more detailed caption is needed. Figures should be independent of article text, so maybe a step by step caption would be useful. Currently I can guess what the figure is saying, but a detailed caption would really help.

**[Done]**

> * Latin names for pathogens should be in italics (typically, not 100% on what JOSS' policy is).

**Yes, sure. Sorry, the formatting was lost during the submission.**

> * When you say "contains modules for" do you mean software modules? If so could you state what the changes are and how difficult would it be to say add another disease for another plant. i.e. if I wanted to do this method for rice plants is it the case of 1 week, 1 month, 1 year? If your method can be applied more widely then it would be to your benefit to state this. Possibly just a sentence on "This method was developed for mildew, with little modification it works with wheat, thus we propose that many other diseases could be applied without difficulty", just an example.

**We mean software modules, though hardware changes might be required as well. We cannot give a clear statement of how much effort will take adding a module for a new disease since the appearance of the disease symptoms and thus the segmentation approach can differ a lot. One day to one year is a pretty realistic interval.**

> * I'd be careful of your conclusion statements. "... Is essential to truly understand the defense mechanism", from reading this text you haven't explicitly stated how your method provides information on defence mechanisms, just that you capture data. Sorry, but this is a subtlety that I feel is important to expand upon in the text. It would give a lot more impact if you clearly state that your method is very useful because X.

**Sorry, the Conclusion part was not meant to be part of the final text. It went there by mistake. It is removed now.**

> * There are a few spelling and grammar mistakes I haven't pointed out. I think copying and pasting your text into a word-processor (Libre-office or MS-word) and setting to English spellcheck would help a lot.

**Sure, we do it in MS Word plus an additional spelling checker but even this doesn’t help sometimes. We did our best and hopefully, the spelling is better now.**

> * One extra thing I just noticed before hitting comment. I don't see that you've referenced your figure in the text. Maybe it would be helpful to break down also in the text what the steps for this method are i.e. what materials and steps does someone need to do if they wanted to use this method themselves.

**[Done]**

> Again, I do hope you see that these are comments of my opinion and not always correct. I'm still going through the rest of the repository and will have more comments shortly.

**Thank you very much for your efforts!**
