+++ paper/paper.bib
@@ -34,7 +34,7 @@ eprint = {
 }
 
 @book{2016,
-author = {Stanley, Johnson and Gnanadhas, Preetha},
+author = {Stanley, Johnson and Preetha, Gnanadhas},
 year = {2016},
 month = {01},
 pages = {1-502},
@@ -98,7 +98,7 @@ title = {Wheat Diseases and Pests: A Guide for Field Identification (2nd Edition
 }
 
 @article{2012_2,
-author = {Dean, RALPH and VAN KAN, JAN A. L. and PRETORIUS, ZACHARIAS A. and HAMMOND-KOSACK, KIM E. and DI PIETRO, ANTONIO and SPANU, PIETRO D. and RUDD, JASON J. and DICKMAN, MARTY and KAHMANN, REGINE and ELLIS, JEFF and FOSTER, GARY D.},
+author = {Dean, R., Van Kan, J.A.L., Pretorius, Z.A., Hammond-Kosack, K.E., Di Pietro, A., Spanu, P.D., Rudd, J.J., Dickman, M., Kahmann, R., Ellis, J. and Foster, G.D.},
 title = {The Top 10 fungal pathogens in molecular plant pathology},
 journal = {Molecular Plant Pathology},
 volume = {13},
