+++ paper/paper.bib
@@ -129,7 +129,7 @@ eprint = {
         https://aip.scitation.org/doi/pdf/10.1109/MCSE.2011.37}}
 	
 @article{2011_2,
-author = {Huerta and Singh, Ravi and Germán, Silvia and McAllum, B.D. and Park, R.F. and W.Q.Chen, and Bhardwaj, S.C. and Goyeau, and},
+author = {Huerta-Espino J, Singh R, German S, McCallum B, Park R, Chen WQ, Bhardwaj S, Goyeau H},
 year = {2011},
 month = {01},
 pages = {143-160},
