+++ paper/paper.bib
@@ -98,7 +98,7 @@ title = {Wheat Diseases and Pests: A Guide for Field Identification (2nd Edition
 }
 
 @article{2012_2,
-author = {Dean, R., Van Kan, J.A.L., Pretorius, Z.A., Hammond-Kosack, K.E., Di Pietro, A., Spanu, P.D., Rudd, J.J., Dickman, M., Kahmann, R., Ellis, J. and Foster, G.D.},
+author = {Dean, R. and Van Kan, J.A.L. and Pretorius, Z.A. and Hammond-Kosack, K.E. and Di Pietro, A. and Spanu, P.D. and Rudd, J.J. and Dickman, M. and Kahmann, R. and Ellis, J. and Foster, G.D.},
 title = {The Top 10 fungal pathogens in molecular plant pathology},
 journal = {Molecular Plant Pathology},
 volume = {13},
