+++ paper/paper.bib
@@ -175,6 +175,17 @@ eprint = {
     abstract = { The emerging picture of plant and animal domestication and agricultural origins in the Near East is dramatically different from that drawn 16 years ago in a landmark article by Bar-Yosef and Meadow. While in 1995 there appeared to have been at least a 1,500-year gap between plant and animal domestication, it now seems that both occurred at roughly the same time, with initial management of morphologically wild future plant and animal domesticates reaching back to at least 11,500 cal BP, if not earlier. A focus on the southern Levant as the core area for crop domestication and diffusion has been replaced by a more pluralistic view that sees domestication of various crops and livestock occurring, sometimes multiple times in the same species, across the entire region. Morphological change can no longer be held to be a leading-edge indicator of domestication. Instead, it appears that a long period of increasingly intensive human management preceded the manifestation of archaeologically detectable morphological change in managed crops and livestock. Agriculture in the Near East arose in the context of broad-based systematic human efforts at modifying local environments and biotic communities to encourage plant and animal resources of economic interest. This process took place across the entire Fertile Crescent during a period of dramatic post-Pleistocene climate and environmental change with considerable regional variation in the scope and intensity of these activities as well as in the range of resources being manipulated. }
 }
 
+@article{2011_5,
+author = {Wellings, Colin},
+year = {2011},
+month = {05},
+pages = {129-141},
+title = {Global status of stripe rust: A review of historical and current threats},
+volume = {179},
+journal = {Euphytica},
+doi = {10.1007/s10681-011-0360-y}
+}
+
 @article{2009,
 author = {Hazell, Peter},
 year = {2009},
