+++ paper/paper.bib
@@ -129,7 +129,7 @@ eprint = {
         https://aip.scitation.org/doi/pdf/10.1109/MCSE.2011.37}}
 	
 @article{2011_2,
-author = {Huerta-Espino J and Singh R and German S and McCallum B and Park R and Chen WQ and Bhardwaj S and Goyeau H},
+author = {Huerta-Espino, J and Singh, R and German, S and McCallum, B and Park, R and Chen, WQ and Bhardwaj, S and Goyeau, H},
 year = {2011},
 month = {01},
 pages = {143-160},
