+++ paper/paper.bib
@@ -90,7 +90,7 @@ doi = {10.1007/978-94-017-7752-0}
 }
 
 @book{2012,
-author = {Duveiller, Etienne and Singh, P.K. and Mezzalama, Monica and Singh, Ravi and Dababat, Abdelfattah},
+author = {Duveiller, Etienne and Singh, P K and Mezzalama, Monica and Singh, Ravi and Dababat, Abdelfattah},
 year = {2012},
 month = {01},
 pages = {},
@@ -211,7 +211,7 @@ abstract = "Phenomics is an emerging transdiscipline dedicated to the systematic
 }
 
 @article{2009_3,
-   author = {Gómez, P. and Rodríguez-Hernández, A. M. and Moury, B. and Aranda, M. A.},
+   author = {Gómez, P and Rodríguez-Hernández, A M and Moury, B and Aranda, M A},
    title = {Genetic resistance for the sustainable control of plant virus diseases: breeding, mechanisms and durability},
    journal = {European Journal of Plant Pathology},
    volume = {125},
