+++ paper/paper.bib
@@ -90,7 +90,7 @@ doi = {10.1007/978-94-017-7752-0}
 }
 
 @book{2012,
-author = {Duveiller, Etienne and Singh, P K and Mezzalama, Monica and Singh, Ravi and Dababat, Abdelfattah},
+author = {Duveiller, E. and Singh, P. K. and Mezzalama, M. and Singh, R. and Dababat, A.},
 year = {2012},
 month = {01},
 pages = {},
