### 2020-06-09 02d86e4
added color checker ignore
