+++ macrobot/cli.py
@@ -15,6 +15,7 @@
 
 import os
 import argparse
+import re
 
 from macrobot.puccinia import RustSegmenter
 from macrobot.bgt import BgtSegmenter
@@ -78,10 +79,11 @@ def main():
                 file_results.write('index' + ';' + 'expNr' + ';' + 'barcode' + ';' + 'Plate_ID' + ';' + 'Lane_ID' + ';' + 'Leaf_ID' + ';' + '%_Inf' + '\n')
                 plates = os.listdir(os.path.join(source_path, experiment, dai))
                 for plate in plates:
-                    img_dir = os.path.join(source_path, experiment, dai, plate)
-                    images = [f for f in os.listdir(img_dir) if f.endswith('.tif')]
-                    processor = segmenter_class(images, img_dir, destination_path, store_leaf_path, experiment, dai, file_results)
-                    processor.start_pipeline()
+                    if not re.search('color', plate, re.IGNORECASE):
+                        img_dir = os.path.join(source_path, experiment, dai, plate)
+                        images = [f for f in os.listdir(img_dir) if f.endswith('.tif')]
+                        processor = segmenter_class(images, img_dir, destination_path, store_leaf_path, experiment, dai, file_results)
+                        processor.start_pipeline()
         except NotADirectoryError:
             print ('Skip ' + source_path + experiment + ' because it is not a valid directory.')
 
