### 2020-06-09 079d61b
added color checker ignore
