+++ macrobot/orga.py
@@ -4,7 +4,6 @@ import urllib.request
 import zipfile
 import shutil
 
-
 def download_test_images(DATA_PATH):
     """Download and unzip test image set from DOI 10.5447/ipk/2020/7"""
 
