### 2020-06-12 snowformatics/macrobot#1 @AoifeHughes
I've just had some time to go through the repository.

I haven't been able to run the tests as it seems dependent on data which is missing (see error below). Is it feasible to possibly include a few images in the repository that would allow potential users to test that the setup works and gives expected results?

```
(macrobot) ➜  macrobot git:(master) pytest
===================================================================================================================== test session starts ======================================================================================================================
platform darwin -- Python 3.8.3, pytest-5.4.3, py-1.8.1, pluggy-0.13.1
rootdir: /Users/nathanhughes/Git/macrobot
collected 3 items / 1 error / 2 selected

============================================================================================================================ ERRORS ============================================================================================================================
________________________________________________________________________________________________________ ERROR collecting macrobot/tests/test_mildew.py ________________________________________________________________________________________________________
macrobot/tests/test_mildew.py:49: in <module>
    plate_id, numer_of_lanes, final_image_list, file_name = mildew_pipeline()
macrobot/tests/test_mildew.py:17: in mildew_pipeline
    experiments = os.listdir(source_path)
E   FileNotFoundError: [Errno 2] No such file or directory: '/Users/nathanhughes/Git/macrobot/macrobot/data'
=================================================================================================================== short test summary info ====================================================================================================================
ERROR macrobot/tests/test_mildew.py - FileNotFoundError: [Errno 2] No such file or directory: '/Users/nathanhughes/Git/macrobot/macrobot/data'
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
====================================================================================================================== 1 error in 16.60s =======================================================================================================================

```
