### 2020-06-15 snowformatics/macrobot#1 @snowformatics
Hi Nathan,

thanks for the comment. Could you please try the test_image example which is explained here:
https://macrobot.readthedocs.io/en/latest/use_case.html

This should automatically download a test image set. After that it should be possible to run all tests.

I have to include the download within the tests, thanks for pointing out.
