### 2020-06-23 #2259 @snowformatics
Dear reviewers @phckopper, @jizecn, @SirSharpest

I uploaded an image test datasets which could be downloaded manually and used with the software. Please unzip the file and use the data folder as main entry for the software, e.g.
`-s C:\Users\name\Desktop\data\ -d C:\Users\name\Desktop\out\ -p mildew`

The folder structure should look like this (it's a bit complex because we use the software in batch mode):
https://github.com/snowformatics/macrobot/blob/master/docs/images/folder_structure.png

As jizecn pointed out, Python Version should be 3.8. I will update the documentation.

I hope it will be easier to test the software now. Please let me know if you have any problems.

Best regards
Stefanie
