### 2020-06-29 snowformatics/macrobot#1 @AoifeHughes
Hi,

Sorry for the slow responses - I've had a paper deadline myself that's eaten all of my time.

I've downloaded the dataset from the link you provided. However, it seems that this isn't enough, is there perhaps something specific that needs done? If so this needs streamlined. All tests should be passable from a simple script without modification

The output from running tests, with the images in the "data" folder.

```
(macrobot) ➜  tests git:(master) ✗ pytest
============================================================================================================================ test session starts =============================================================================================================================
platform darwin -- Python 3.8.3, pytest-5.4.3, py-1.9.0, pluggy-0.13.1
rootdir: /Users/nathan/Git/macrobot/macrobot/tests
collected 3 items / 1 error / 2 selected

=================================================================================================================================== ERRORS ===================================================================================================================================
______________________________________________________________________________________________________________________ ERROR collecting test_mildew.py _______________________________________________________________________________________________________________________
test_mildew.py:49: in <module>
    plate_id, numer_of_lanes, final_image_list, file_name = mildew_pipeline()
test_mildew.py:45: in mildew_pipeline
    return plate_id, numer_of_lanes, final_image_list, file_name
E   UnboundLocalError: local variable 'plate_id' referenced before assignment
------------------------------------------------------------------------------------------------------------------------------ Captured stdout -------------------------------------------------------------------------------------------------------------------------------

=== Start Macrobot pipeline ===
 Experiment: Image test set for Macrobot software validation
Skip /Users/nathan/Git/macrobot/macrobot/dataImage test set for Macrobot software validation because it is not a directory.

=== End Macrobot pipeline ===
========================================================================================================================== short test summary info ===========================================================================================================================
ERROR test_mildew.py - UnboundLocalError: local variable 'plate_id' referenced before assignment
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
============================================================================================================================== 1 error in 0.58s ==============================================================================================================================
```

I've also tried running the command-line app and I get this:

```
(macrobot) ➜  macrobot git:(master) ✗ mb -s data/ -d results/ -p mildew
/Users/nathan/opt/anaconda3/envs/macrobot/lib/python3.8/site-packages/macrobot/data
/Users/nathan/opt/anaconda3/envs/macrobot/lib/python3.8/site-packages/macrobot/data
data/
Namespace(destination_path='results/', procedure='mildew', source_path='data/') <class 'macrobot.bgt.BgtSegmenter'>

=== Start Macrobot pipeline ===
 Experiment: gb2_exp40
Skip data/gb2_exp40 because it is not a valid directory.

=== End Macrobot pipeline ===
Skip data/.DS_Store because it is not a valid directory.

=== End Macrobot pipeline ===
```

This is the file structure I've used: [https://i.imgur.com/DfPLN53.png](https://i.imgur.com/DfPLN53.png)
