### 2020-06-30 snowformatics/macrobot#1 @snowformatics
Hi Nathan,

no problem, I think we are all very busy at the moment..

Thanks for testing again, the folder structure looks correct. Both error messages show that the software has problem with the folder structure. I tried to use os.join() to deal with the file paths but I have no experience how it will work on Linux or Mac. I developed and tried the software only on Windows because our institution and our users supports / use only Windows OS (sorry about that). Would it be possible to try the software on a Windows system? Or maybe the other reviewers could test on Windows?

I am sorry about your inconvenience but as mentioned in the paper, the target of the software is Windows OS.

Best regards
Stefanie
