### 2020-06-30 snowformatics/macrobot#1 @snowformatics
On last try, could you start the command-line app with the full paths please?

So instead of:
mb -s data/ -d results/ -p mildew

This (with linux path):
mb -s C:\Users\xy\Desktop\data\ -d C:\Users\xy\Desktop\results\ -p mildew

Thanks
