### 2020-07-03 #2259 @snowformatics
> @snowformatics I can set the archive once the review is complete. It looks like we are just waiting to finish the issue where @SirSharpest is conducting their part of the review. Once that is done I’ll set the DOI and work through the last steps of the review.

Sorry for the command, I just wanted to  note down the link, no rush...
