### 2020-07-03 #2259 @SirSharpest
@mjsottile My review is complete too. I can recommend for publication.

(My one minor gripe, which I can understand letting go, is software tests should run out of the box without a user needing to source additional data)
