### 2020-07-05 #2259 @snowformatics
Thanks everybody for taking the time to review our project!

@SirSharpest I will reconsider my tests for future projects in order to exclude integration tests with real world images. Thanks for the suggestion.
