### 2020-07-05 snowformatics/macrobot#12 @mjsottile
JOSS paper review
I found a few minor things to address in the paper before final acceptance.  They are:

- The space in ">10 000 years" causes the number 10000 to be split across a line break in an awkward way.  I would suggest revising to avoid having the number span a line break, such as removing the space in the number.

- The wording in the last paragraph of the summary has some minor bugs.  For example, I would change:

> "...and will work without major modification also for barley leaf and stripe rusts and probably some other leaf disease with a similar appearance.  Phenotyping of other disease and non-disease related phenotypes will require developments of dedicated modules."

to:

> "...and will work without major modification for barley leaf, stripe rusts and probably other leaf diseases with a similar appearance.  Phenotyping of other disease and non-disease related phenotypes will require development of dedicated modules."

- Final paragraph, change the following to address grammar and spelling:

> "...and was build and tested in Microsoft Windows enviroment"

to:

> "...and was built and tested in the Microsoft Windows environment"
