### 2020-07-05 #2259 @mjsottile
It looks like the reviews are complete and we can proceed to final steps.  I did a pass over the paper and found a few minor issues that should take very little time to address.  They are listed in the project issue https://github.com/snowformatics/macrobot/issues/12.

Once you've addressed those I will submit the paper for final processing by the EIC team.
