### 2020-07-06 4fb1fc3
Update paper.md
