### 2020-07-06 61ace4f
Update paper.md
