### 2020-07-06 snowformatics/macrobot#12 @snowformatics
Thanks for your suggestions @mjsottile ! I corrected everything and hope it's ok now.
