### 2020-07-06 #2259 @mjsottile
@snowformatics Looks like one final issue with one of the DOIs.  Can you fix that so I can issue the accept command?  I didn’t notice that when I did my final review of the paper earlier.
