+++ paper/paper.bib
@@ -146,7 +146,7 @@ number = "8",
 pages = "1536 - 1545",
 year = "2011",
 issn = "0921-8009",
-doi = "https://doi.org/10.1016/j.ecolecon.2011.03.016",
+doi = "10.1016/j.ecolecon.2011.03.016",
 url = "http://www.sciencedirect.com/science/article/pii/S0921800911001054",
 author = "Naoufel Mzoughi",
 keywords = "Integrated protection, Organic farming, Social and moral concerns",
