### 2020-07-09 3bc47a4
Update paper.md
