+++ paper/paper.bib
@@ -72,7 +72,7 @@ doi = {10.1007/978-94-017-7752-0}
    DOI = {10.1186/s13059-014-0518-8},
    url = {https://doi.org/10.1186/s13059-014-0518-8},
    year = {2014},
-   type = {Journal Article}
+
 }
 
 @article{2013,
@@ -86,7 +86,7 @@ doi = {10.1007/978-94-017-7752-0}
    DOI = {10.1007/s13593-012-0105-x},
    url = {https://doi.org/10.1007/s13593-012-0105-x},
    year = {2013},
-   type = {Journal Article}
+
 }
 
 @book{2012,
@@ -135,6 +135,7 @@ month = {01},
 pages = {143-160},
 title = {Global status of wheat leaf rust caused by Puccinia triticina},
 volume = {179},
+doi = {10.1007/s10681-011-0361-x},
 journal = {Euphytica}
 }
 
@@ -221,7 +222,7 @@ abstract = "Phenomics is an emerging transdiscipline dedicated to the systematic
    DOI = {10.1007/s10658-009-9468-5},
    url = {https://doi.org/10.1007/s10658-009-9468-5},
    year = {2009},
-   type = {Journal Article}
+  
 }
 
 @article{2007,
@@ -234,7 +235,6 @@ abstract = "Phenomics is an emerging transdiscipline dedicated to the systematic
    ISSN = {1524-0215 (Print)
 1524-0215},
    year = {2007},
-   type = {Journal Article}
 }
 
 @article{2005,
