+++ paper/paper.bib
@@ -193,6 +193,11 @@ year = {2009},
 month = {01},
 pages = {},
 title = {The Asian Green Revolution}
+institution = {International Food Policy Research Institute (IFPRI)},
+type = {IFPRI discussion papers},
+number = {911},
+keywords = {millions fed; Green Revolution; agricultural transformation},
+url = {https://EconPapers.repec.org/RePEc:fpr:ifprid:911}
 }
 
 @article{2009_2,
