+++ paper/paper.bib
@@ -192,7 +192,7 @@ author = {Hazell, Peter},
 year = {2009},
 month = {01},
 pages = {},
-title = {The Asian Green Revolution}
+title = {The Asian Green Revolution},
 institution = {International Food Policy Research Institute (IFPRI)},
 type = {IFPRI discussion papers},
 number = {911},
