### 2020-07-09 #2259 @snowformatics
> Hi @snowformatics, just a few minor edits remaining on the paper before we accept:
>
> * In the first paragraph (and anywhere else this occurs), when you have multiple citations please combine them like this:
>   `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
> * the Hazell reference is missing details
> * please add the DOI 10.1007/s10681-011-0361-x to the Huerta-Espino article
> * a few references include "journal article" in the citations, please remove that where it appears

Thanks for your comment, I hope everything is ok now.
