### 2020-07-09 #2259 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
